fetch indicators

+++ server/service/src/demographic/insert_demographic_projection.rs
@@ -22,11 +22,11 @@ pub enum InsertDemographicProjectionError {
 pub struct InsertDemographicProjection {
     pub id: String,
     pub base_year: i32,
-    pub year_1: Option<i32>,
-    pub year_2: Option<i32>,
-    pub year_3: Option<i32>,
-    pub year_4: Option<i32>,
-    pub year_5: Option<i32>,
+    pub year_1: Option<f64>,
+    pub year_2: Option<f64>,
+    pub year_3: Option<f64>,
+    pub year_4: Option<f64>,
+    pub year_5: Option<f64>,
 }
 
 pub fn insert_demographic_projection(
